Authentication

+++ src/routers/task.router.js
@@ -1,66 +1,71 @@
 const express = require('express');
 const router = express.Router();
 const Task = require('../models/task.model');
+const auth = require('../middleware/auth.middleware');
 
-router.post('/tasks', (req, res)=>{
-    const task = new Task(req.body);
-    
-    task.save().then((task)=>{
+//creating a brand new task - endpoint 
+router.post('/tasks', auth, (req, res) => {
+    const task = new Task({
+        ...req.body,
+        owner: req.user._id
+    });
+
+    task.save().then((task) => {
         res.status(201).send(task);
-    }).catch(error=>{
+    }).catch(error => {
         res.status(400).send(error);
     })
 })
 
-router.get('/tasks', (req, res)=>{
-    Task.find({}).then(tasks => {
-        res.status(200).send(tasks);
-    }).catch(error=>{
-        res.status(500).send(error);
-    })
+//getting all tasks associated with current user - endpoint
+router.get('/tasks', auth, async (req, res) => {
+    await req.user.populate('tasks').execPopulate();
+    res.send(req.user.tasks);
 })
 
-router.get('/tasks/:id', (req, res)=>{
-    const _id= req.params.id;
-    Task.find({_id}).then(user => {
-        res.status(200).send(user);
-    }).catch(error=>{
-        if(error.kind=="ObjectId"){
-            return res.status(404).send('No user found with matching ID');
-        }
-        res.status(500).send(error);
-    })
+//getting a task by id - endpoint
+router.get('/tasks/:id', auth, async (req, res) => {
+    const _id = req.params.id;
+    const task = await Task.findOne({ _id, owner: req.user._id })
+    if (!task) {
+        res.status(404).send({ error: "task not found" });
+    }
+    res.status(200).send(task);
 })
 
-router.patch('/tasks/:id', async (req, res)=>{
-    const _id= req.params.id;
+//updating a task - endpoint
+router.patch('/tasks/:id', auth, async (req, res) => {
+    const _id = req.params.id;
     const allowedUpdates = ['description', 'completed'];
-
-    const isValidUpdate = Object.keys(req.body).every(update=>allowedUpdates.includes(update));
-    if(!isValidUpdate){
+    const updates = Object.keys(req.body);
+    const isValidUpdate = updates.every(update => allowedUpdates.includes(update));
+    if (!isValidUpdate) {
         return res.status(400).send('Unrecognized fields to update');
     }
 
-    try{
-        const task = await Task.findByIdAndUpdate({_id}, req.body, { new: true, runValidators: true});
-        if(!task){
-            return res.status(404).send('task not found with matching ID');
+    try {
+        const task = await Task.findOne({ _id, owner: req.user._id });
+        if (!task) {
+            return res.status(404).send('task not found');
         }
+        updates.forEach(update => task[update] = req.body[update]);
+        await task.save();
         res.status(202).send(task);
-    }catch(error){
+    } catch (error) {
         res.status(400).send(error);
     }
 })
 
-router.delete('/tasks/:id', async (req,res)=>{
-    const _id= req.params.id;
-    try{
-        const task = await Task.findByIdAndDelete({_id});
-        if(!task){
-            return res.status(404).send("task not found with matching ID")
+//deleting a task - endpoint
+router.delete('/tasks/:id', auth, async (req, res) => {
+    const _id = req.params.id;
+    try {
+        const task = await Task.findByIdAndDelete({ _id, owner: req.user._id });
+        if (!task) {
+            return res.status(404).send({error: "task not found"})
         }
-        res.status(202).send(task);
-    }catch(error){
+        res.status(200).send(task);
+    } catch (error) {
         res.status(500).send(error);
     }
 });
